Add imperative functions to overwrite krassenstein fields
When I set the `textContent` of a node to `krassenstein`, I also set
the node's `data-krassenstein-original-text-content` field to the
actual content (i.e., `sputnikhaslande`).

Then, when I want to dekrassensteinify the page, I can just select all
nodes with a `data-krassenstein-original-text-content` field and set
that value to the node's textContent.